Add the Scope section
Add the Scope section to clarify In Scope and Out of scope.

And also replace "retarget" with "remap" because "retarget" may be
often used for applying the same animation among different structured
skeletons but it is out of scope in this extension.

The scope of this extension is applying the same animation among
the same or similar structured skeletons. Using "retarget" for
it doesn't seem to be really wrong but "remap" may be less confusing.