Add left/rightEyes and jaw bones
Currently default skeleton predefined in the extension is based on
the one in VRM. And VRM skeleton is based on the Unity one.

I first thought that left/rightEyes and jaw bones might not be
needed in major humanoid models because facial animation might
be applied in different way (ex: morph) but there may be no
reason to follow the existing major definition. Probably adding
them doesn't hurt.